Add missing modules to PREREQ_PM
Even though the distribution installs and builds correctly, mentioning all
required modules helps things like metacpan find and display the correct
dependencies.  This helps the dist's kwalitee and aids in determining the
"river of CPAN" (see http://neilb.org/2015/04/20/river-of-cpan.html), which
is good for the entire Perl module ecosystem.