Specify an explicit minimum Perl version
I know we already discussed this on IRC, but I happened to find that the
minimum version was actually higher than 5.6.  It turns out that certain
minimum syntax features used in Template::Flute require at least 5.8, thus
it would help anyone installing the module on an old system (and anyone
testing on old systems) if they knew early on that the module won't work
without at least this minimum version.  Hence, I've added this commit here.
It's also not a large change, so I hope you can accept it.