Use full package name for test functions not imported into test script scope
It turns out the workaround I was using to get test functions such as
`ok_manifest` and `all_pod_coverage_ok` imported into the current scope so
that when `RELEASE_TESTING` was set, the modules would be loaded and the
correct functions run.  Unfortunately, this was incorrect.  All `use`
statements are run at script compile time (they run in effectively a `BEGIN`
block).  Unfortunately, my testing didn't check the case when the modules
were not locally installed; when the relevant modules (e.g.
Test::CheckManifest) are installed, then everything works fine, both with and
without `RELEASE_TESTING`.  Fortunately, upon working on another pull
request, Travis showed up the issue.

The workaround I'm now using -- since `ok_manifest` is not seen in the
`main` scope (due to `try_load_class` effectively running `require` and not
`use`) -- is to use the complete package path to the relevant test
functions.  It's not pretty, but it works in the cases I mention above,
namely with and without `RELEASE_TESTING` and with and without the required
modules being installed.

My apologies for the hassle!